shipment update completed

+++ courierwala-server/src/main/java/com/courierwala/server/entities/CourierOrder.java
@@ -55,7 +55,7 @@ public class CourierOrder extends BaseEntity {
     private Address deliveryAddress;
 
 
-    private BigDecimal packageWeight;
+    private double packageWeight;
     private String packageType;
 
     @Enumerated(EnumType.STRING)
